NewsPortal/requirements.txt
django
django_import_export
django-filter
django-apscheduler
celery
django-modeltranslation
django-allauth
